requirements.txt
gradio==6.0.1
numpy==2.3.5
opencv_python==4.9.0.80
Pillow==12.0.0
torch==1.13.0
torchvision==0.14.0
